refactor(resource): mv package from pkg/ to internal/
