+++ cmd/main.go
@@ -40,7 +40,7 @@ import (
 	resourcev1 "github.com/antimetal/agent/pkg/api/resource/v1"
 	"github.com/antimetal/agent/pkg/performance"
 	_ "github.com/antimetal/agent/pkg/performance/collectors" // Register collectors
-	"github.com/antimetal/agent/pkg/resource/store"
+	"github.com/antimetal/agent/internal/resource/store"
 )
 
 var (
