+++ internal/hardware/graph/builder_test.go
@@ -15,8 +15,8 @@ import (
 	hardwarev1 "github.com/antimetal/agent/pkg/api/antimetal/hardware/v1"
 	resourcev1 "github.com/antimetal/agent/pkg/api/resource/v1"
 	"github.com/antimetal/agent/pkg/performance"
-	"github.com/antimetal/agent/pkg/resource"
-	"github.com/antimetal/agent/pkg/resource/store"
+	"github.com/antimetal/agent/internal/resource"
+	"github.com/antimetal/agent/internal/resource/store"
 	"github.com/go-logr/logr"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
