+++ internal/kubernetes/agent/controller.go
@@ -13,7 +13,7 @@ import (
 	"time"
 
 	"github.com/antimetal/agent/pkg/errors"
-	"github.com/antimetal/agent/pkg/resource"
+	"github.com/antimetal/agent/internal/resource"
 	"github.com/go-logr/logr"
 	gogoproto "github.com/gogo/protobuf/proto"
 	"golang.org/x/sync/errgroup"
