+++ internal/resource/store/encode.go
