+++ internal/resource/store/encode_test.go
