+++ internal/resource/types.go
